rdp: Implement ringbuffer to feed the displaylist
This improves performance as the RDP will start rendering when the first
command is written, thus working in parallel with the CPU.

The concept of reusing displaylist and tracking RDP state is broken,
since the RDP state must be known at creation time of the displaylist.
There aren't many use cases for static displaylists anyway, so now there
is only one global display list.