periph: Use Pool for Notes and fix race in dma()
The Pool is necessary because we don't know how long the note will be
retained by the caller of dma().  It could cause a situation where the
note is reused and then multiple goroutines sleeping on it.

The race condition was caused by pushing a job into the queue.  If we
get interrupted before acquiring dmaActive, the queue might already be
empty, or we Pop a different job from it, returning an invalid note.